# housingfund-loan/loan-swagger.yaml
...
swagger: '2.0'
info:
  title: Handge housingfund Platform API
  description: 提供公积金的数据支持
  termsOfService: Handge.
  contact:
    name: API Documentation Support. QQ:17213069
    url: http://handge.cn
    email: front_end@handge.cn
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0.1"
# the domain of the service
host: 127.0.0.1:8001
...
schemes:
  - http
# will be prefixed to all paths
basePath: /loan
security:
  - addToken: []
...
produces:
  - application/json
paths:
...
  /applyLoan/{YWLSH}:
    get:
      description: 详情页面
      operationId: getApplyDetails
      tags:
        - housing
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: string
          description: 业务流水号
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/getApplicantResponse'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      description: 修改页面
      operationId: putModifyApplication
      tags:
        - housing
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: string
          description: 业务流水号
        - name: body
          required: true
          in: body
          schema:
            $ref: '#/definitions/ApplicantPost'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: 删除申请
      operationId: deleteApplication
      tags:
        - housing
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: string
          description: 业务流水号
        - name: YWWD
          required: true
          in: query
          type: string
          description: 网点id
        - name: CZY
          required: true
          in: query
          type: string
          description: 操作员
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /signContract:
    post:
      description: 签订贷款合同
      operationId: postPurchaseContract
      tags:
        - housing
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/SignContractPost'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
# <放款>
  /loans/{YWLSH}:
    get:
      description: 发放贷款信息
      operationId: getConfirmLoans
      tags:
        - housing
      parameters:
        - name: YWLSH
          in: path
          required: true
          type: string
          description: 业务流水号
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/confirmPaymentResponse'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      description: 点击确认放款
      operationId: putConfirmLoansSubmit
      tags:
        - housing
      parameters:
        - name: YWLSH
          in: path
          required: true
          type: string
          description: 业务流水号
        - name: body
          required: true
          in: body
          schema:
            $ref: '#/definitions/ConfirmPaymentPut'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /account/search:
    get:
      description: 贷款账户查询
      operationId: getHousingFundAccountSearch
      tags:
        - housing
      parameters:
        - name: DKZH
          in: query
          type: string
          description: 贷款账号
        - name: JKRXM
          in: query
          type: string
          description: 借款人姓名
        - name: status
          in: query
          type: string
          description: 状态（0:正常 1:逾期 2:呆账 3:转出 4:结清）
        - name: pageSize
          in: query
          type: string
          description: 每页数量
        - name: pageNo
          in: query
          type: string
          description: 页码
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/HousingFundAccountGet'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /repayment/state/{YWLSH}/{action}/{role}:
    put:
      description: 提交与撤回还款申请
      operationId: putEstateStatusSubmit
      tags:
        - housing
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: integer
          description: 业务流水号
        - name: action
          required: true
          in: path
          type: string
          description: 状态（0：提交 1：撤回）
        - name: role
          required: true
          in: path
          type: string
          description: 角色
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
#<还款审核列表-还款申请>
...
  /estateProject:
    get:
      description: 楼盘项目列表（申请受理）共用
      operationId: getEstateList
      tags:
        - estateProject
      parameters:
        - name: LPMC
          in: query
          type: string
          description: 楼盘名称
        - name: FKGS
          in: query
          type: string
          description: 房开公司
        - name: pageNo
          in: query
          type: string
          description: 当前页数
        - name: pageSize
          in: query
          type: string
          description: 当前页数数据条数
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/EstateProjectGet'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: 新增楼盘项目
      operationId: postEstate
      tags:
        - estateProject
      parameters:
        - name: action
          in: query
          type: string
          description: 状态（0：保存 1：提交）
        - name: estate
          required: true
          in: body
          schema:
            $ref: '#/definitions/EstateProjectPost'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
# <选择房开公司-楼盘项目>
  /estateProjectAddSearch:
    get:
      description: 选择房开公司
      operationId: getEstate
      tags:
        - estateProject
      parameters:
        - name: FKZH
          in: query
          type: string
          description: 房开账号
        - name: FKGS
          type: string
          in: query
          description: 房开公司
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/estateProjectAddSearchGet'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
# <详情、修改-楼盘项目>
  /estateProject/{YWLSH}:
    get:
      description: 楼盘项目详情（打印回执单）
      operationId: getEstateId
      tags:
        - estateProject
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: string
          description: 业务流水号
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/EstateIdGet'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      description: 修改楼盘项目信息
      operationId: putEstateId
      tags:
        - estateProject
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: string
          description: 业务流水号
        - name: CZLX
          in: query
          type: string
          description: 操作类型
        - name: estate
          required: true
          in: body
          schema:
            $ref: '#/definitions/EstatePut'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: 删除楼盘项目信息
      operationId: deleteEstateId
      tags:
        - estateProject
      parameters:
        - name: YWLSH
          required: true
          in: path
          type: string
          description: 业务流水号
        - name: WD_id
          in: query
          required: true
          type: string
          description: 网点id
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/CommonResponses'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
# <受理楼盘列表-楼盘项目>
...
securityDefinitions:
  addToken:
    type: apiKey
    in: header
    name: token
definitions:
...
  getApplicantResponse:
    properties:
      Res:
        type: object
        properties:
          YWLSH:
            type: string
            description: 业务流水号
          #<申请贷款>
          applicantInformation:
            type: object
            properties:
              JCD:
                type: string
                description: 缴存地
              JKRGJJZH:
                type: string
                description: 借款人公积金账号
              #<借款人信息>
              borrowerInformation:
                type: object
                properties:
                  JKRXM:
                    type: string
                    description: 借款人姓名
                  JKRZJLX:
                    type: string
                    description: 借款人证件类型
                  JKRZJHM:
                    type: string
                    description: 借款人证件号码
                  CSNY:
                    type: string
                    description: 出生年月
                  XingBie:
                    type: string
                    description: 性别
                  NianLing:
                    type: string
                    description: 年龄
                  XueLi:
                    type: string
                    description: 学历
                  HYZK:
                    type: string
                    description: 婚姻状态（0：已婚 1：未婚 2：丧偶 3：离婚 4：未说明的婚姻状况）
                  JKZK:
                    type: string
                    description: 健康状态（0：良好 1：一般 2：差）
                  ZhiCheng:
                    type: string
                    description: 职称
                  ZhiWu:
                    type: string
                    description: 职务
                  YGXZ:
                    type: string
                    description: 用工性质（0：正式职工 1：合同制 2：聘用制）
                  ZYJJLY:
                    type: string
                    description: 主要经济来源（0：工资收入 1：个体经营收入 2：其他非工资收入）
                  YSR:
                    type: string
                    description: 月收入
                  JTYSR:
                    type: string
                    description: 家庭月收入
                  GDDHHM:
                    type: string
                    description: 固定电话号码
                  SJHM:
                    type: string
                    description: 手机号码
                  JTZZ:
                    type: string
                    description: 家庭住址
                  HKSZD:
                    type: string
                    description: 户口所在地
              #<公积金账号信息>
              accountInformation:
                type: object
                properties:
                  JZNY:
                    type: string
                    description: 缴至年月
                  GRZHZT:
                    type: string
                    description: 个人账户状态
                  YJCE:
                    type: string
                    description: 月缴存额
                  GRJCJS:
                    type: string
                    description: 个人缴存基数
                  GRZHYE:
                    type: string
                    description: 个人账户余额
              #<单位信息>
              unitInformation:
                type: object
                properties:
                  DWMC:
                    type: string
                    description: 单位名称
                  DWZH:
                    type: string
                    description: 单位账号
                  DWDZ:
                    type: string
                    description: 单位地址
                  DWDH:
                    type: string
                    description: 单位电话
                  DWLB:
                    type: string
                    description: 单位性质
              SCZL:
                type: string
                description: 上传上传资料
          #<共同借款人>
          commonBorrowerInformation:
            type: object
            properties:
              CHGX:
                type: string
                description: 参货关系 （0：本人或户主 1：配偶 2：子 3：女 4：孙子、孙女或外孙子、外孙女 5：父母 6：祖父母或外祖父母 7：兄、弟、姐、妹 8：其他）
              GTJKRGJJZH:
                type: string
                description: 共同借款人公积金账号
              JCD:
                type: string
                description: 缴存地
              GTJKRXM:
                type: string
                description: 共同借款人姓名
              GTJKRZJLX:
                type: string
                description: 共同借款人证件类型
              GTJKRZJHM:
                type: string
                description: 共同借款人证件号码
              GDDHHM:
                type: string
                description: 固定电话号码
              SJHM:
                type: string
                description: 手机号码
              YSR:
                type: string
                description: 月收入
              SCZL:
                type: string
                description: 上传上传资料
          #<房屋信息>
          houseInformation:
            type: object
            properties:
              DKYT:
                type: string
                description: 贷款用途
              SFWESF:
                type: string
                description: 是否为二手房
              #<购买信息>
              purchaseFirstInformation:
                type: object
                properties:
                  LPMC:
                    type: string
                    description: 楼盘名称
                  DanJia:
                    type: string
                    description: 单价
                  HTJE:
                    type: string
                    description: 合同金额
                  FWZJ:
                    type: string
                    description: 房屋总价
                  GFHTBH:
                    type: string
                    description: 购房合同编号
                  FWJZMJ:
                    type: string
                    description: 房屋建筑面积
                  FWTNMJ:
                    type: string
                    description: 房屋套内面积
                  SFK:
                    type: string
                    description: 首付款
                  FWXZ:
                    type: string
                    description: 房屋性质（0：商品房 1：障碍性住房 2：自建房 3：两限房 4：集资房 5：危改房 6：经济适用房 7：房改房 8：其他）
                  FWXS:
                    type: string
                    description: 房屋形式（0：在建房  1：现房）
                  FWJG:
                    type: string
                    description: 房屋结构（0：框架 1：砖混 2：其他）
                  FWJGRQ:
                    type: string
                    description: 房屋竣工日期
                  FWZL:
                    type: string
                    description: 房屋坐落
                  SFRMC:
                    type: string
                    description: 售房人名称
                  LXFS:
                    type: string
                    description: 联系方式
                  SPFYSXKBH:
                    type: string
                    description: 商品房预售许可编号
                  SFRZHHM:
                    type: string
                    description: 售房人账户号码
                  SFRKHYHMC:
                    type: string
                    description: 售房人开户银行名称
                  SFRYHMC:
                    type: string
                    description: 售房人银行名称
              purchaseSecondInformation:
                type: object
                properties:
                  DanJia:
                    type: string
                    description: 单价
                  HTJE:
                    type: string
                    description: 合同金额
                  FWZJ:
                    type: string
                    description: 房屋总价
                  GFHTBH:
                    type: string
                    description: 购房合同编号
                  FWJZMJ:
                    type: string
                    description: 房屋建筑面积
                  FWTNMJ:
                    type: string
                    description: 房屋套内面积
                  SFK:
                    type: string
                    description: 首付款
                  FWXZ:
                    type: string
                    description: 房屋性质（0：商品房 1：障碍性住房 2：自建房 3：两限房 4：集资房 5：危改房 6：经济适用房 7：房改房 8：其他）
                  FWXS:
                    type: string
                    description: 房屋形式（0：在建房  1：现房）
                  FWJG:
                    type: string
                    description: 房屋结构（0：框架 1：砖混 2：其他）
                  FWJGRQ:
                    type: string
                    description: 房屋竣工日期
                  FWZL:
                    type: string
                    description: 房屋坐落
                  SFRMC:
                    type: string
                    description: 售房人名称
                  LXFS:
                    type: string
                    description: 联系方式
                  GRSKYHZH:
                    type: string
                    description: 个人收款银行账号
                  KHYHMC:
                    type: string
                    description: 开户银行名称
                  YHKHM:
                    type: string
                    description: 银行开户名
              #<自建、翻修信息>
              buildInformation:
                type: object
                properties:
                  PZJGWH:
                    type: string
                    description: 批准机关文号
                  JHJZFY:
                    type: string
                    description: 计划建造费用
                  JZCS:
                    type: string
                    description: 建造层数
                  JHKGRQ:
                    type: string
                    description: 计划开工日期
                  JHJGRQ:
                    type: string
                    description: 计划竣工日期
                  TDSYZH:
                    type: string
                    description: 土地使用证号
                  JZGCGHXKZH:
                    type: string
                    description: 建造工程规划许可证号
                  JZYDGHXKZH:
                    type: string
                    description: 建造用地规划许可证号
                  GRSYZJ:
                    type: string
                    description: 个人使用资金
                  FWZLDZ:
                    type: string
                    description: 房屋坐落地址
                  JCHJZMJ:
                    type: string
                    description: 建成后建造面积
                  FWJG:
                    type: string
                    description: 房屋结构（0：框架 1：砖混 2：其他）
                  FWXZ:
                    type: string
                    description: 房屋性质（0：商品房 1：障碍性住房 2：自建房 3：两限房 4：集资房 5：危改房 6：经济适用房 7：房改房 8：其他）
                  GRSKYHZH:
                    type: string
                    description: 个人收款银行账号
                  KHHYHMC:
                    type: string
                    description: 开户行银行名称
                  YHKHM:
                    type: string
                    description: 银行开户名
                  SCZL:
                    type: string
                    description: 上传上传资料
              #<大修信息>
              overhaulInformation:
                type: object
                properties:
                  PZJGWH:
                    type: string
                    description: 批准机关文号
                  JHFXFY:
                    type: string
                    description: 计划翻修费用
                  FWXZ:
                    type: string
                    description: 房屋性质（0：商品房 1：障碍性住房 2：自建房 3：两限房 4：集资房 5：危改房 6：经济适用房 7：房改房 8：其他）
                  FXHCS:
                    type: string
                    description: 翻修后层数
                  JHKGRQ:
                    type: string
                    description: 计划开工日期
                  JHJGRQ:
                    type: string
                    description: 计划竣工日期
                  YJZMJ:
                    type: string
                    description: 原建筑面积
                  FWZLDZ:
                    type: string
                    description: 房屋坐落地址
                  GRSKYHZH:
                    type: string
                    description: 个人收款银行账号
                  KHHYHMC:
                    type: string
                    description: 开户行银行名称
                  YHKHM:
                    type: string
                    description: 银行开户名
                  SCZL:
                    type: string
                    description: 上传上传资料
          #<资金信息>
          capitalInformation:
            type: object
            properties:
              DKLX:
                type: string
                description: 贷款类型（0：公积金贷款 1：组合贷款 2：贴息贷款 3：其他）
              DKDBLX:
                type: string
                description: 贷款担保类型  （0：抵押 1：质押 2：保证 3：其他）
              HKFS:
                type: string
                description: 还款方式（0：等额本息 1：等额本金 2：一次还款付息 3：自由还款方式 4：其他）
              HTDKJE:
                type: string
                description: 合同贷款金额
              HTDKJEDX:
                type: string
                description: 合同贷款金额大写
              DKQS:
                type: string
                description: 贷款期数
              FWTS:
                type: string
                description: 房屋套数（0：一套 1：二套 2：三套 3：四套  5：五套及以上）
              JKHTLL:
                type: string
                description: 借款合同利率
              LLFSBL:
                type: string
                description: 利率浮动比例
              ZXLL:
                type: string
                description: 执行利率
              SYDKQS:
                type: string
                description: 商业贷款期数
              SYDKJE:
                type: string
                description: 商业贷款金额
              SYDKYH:
                type: string
                description: 商业贷款银行（0：无 1：中国银行 2：建设银行 3：农业银行 4：工商银行 5：交通银行）
              SYDKYHKE:
                type: string
                description: 商业贷款月还款额
          #<担保信息>
          collateralInformation:
            type: object
            properties:
              DKDBLX:
                type: string
                description: 贷款担保类型
              #<抵押>
              mortgageInformation:
                type: object
                properties:
                  DYWSYQRXM:
                    type: string
                    description: 抵押物所有权人姓名
                  DYWSYQRSFZHM:
                    type: string
                    description: 抵押物所有权人身份证号码
                  DYWSYQRLXDH:
                    type: string
                    description: 抵押物所有权人联系电话
                  DYWGYQRSFZHM:
                    type: string
                    description: 抵押物共有权人身份证号码
                  DYWGYQRLXDH:
                    type: string
                    description: 抵押物共有权人联系电话
                  DYWMC:
                    type: string
                    description: 抵押物名称
                  QSZSBH:
                    type: string
                    description: 权属证书编号
                  DYWPGJZ:
                    type: string
                    description: 抵押物评估价值
                  DYFWXS:
                    type: string
                    description: 抵押房屋形式（0：住宅 （期房） 1：住宅（现房） 2：商铺 3：其他）
                  FWMJ:
                    type: string
                    description: 房屋面积1
                  FWJG:
                    type: string
                    description: 房屋结构（0：框架 1：砖混 2：土木 3：其他）
                  DYWFWZL:
                    type: string
                    description: 抵押物房屋坐落
              #<质押>
              pledgeInformation:
                type: object
                properties:
                  ZYWSYQRXM:
                    type: string
                    description: 质押物所有权人姓名
                  ZYWSYQRSFZHM:
                    type: string
                    description: 质押物所有权人身份证号码
                  ZYWSYQRLXDH:
                    type: string
                    description: 质押物所有权人联系电话
                  ZYWMC:
                    type: string
                    description: 质押物名称
                  ZYWJZ:
                    type: string
                    description: 质押物价值
              #<保证>
              guaranteeInformation:
                type: object
                properties:
                  BZFLX:
                    type: string
                    description: 保证方类型（0：个人 1：机构）
                  BZRXM:
                    type: string
                    description: 保证人姓名
                  BZRSFZHM:
                    type: string
                    description: 保证人身份证号码
                  BZRLXDH:
                    type: string
                    description: 保证人联系电话
                  BZRXJZDZ:
                    type: string
                    description: 保证人现居住地址
                  TXDZ:
                    type: string
                    description: 通讯地址
                  YZBM:
                    type: string
                    description: 邮政编码
              SCZL:
                type: string
                description: 上传上传资料
          #<其他资料>
          SCZL:
            type: string
            description: 上传上传资料
          #<审核信息>
          reviewInformation:
            description:  审核信息
            type: array
            items:
              type: object
              properties:
                YiJian:
                  type: string
                  description: 意见
                CZY:
                  type: string
                  description: 操作员
                ZhiWu:
                  type: string
                  description: 职务
                YWWD:
                  type: string
                  description: 业务网点
                QuDao:
                  type: string
                  description: 渠道
                ShiJian:
                  type: string
                  description: 时间
...
  ConfirmPaymentPut:
    properties:
      Req:
        type: object
        properties:
          YWLSH:
            type: string
            description: 业务流水号
          JKHTBH:
            type: string
            description: 借款合同编号
          DKZH:
            type: string
            description: 贷款账号
          JKRXM:
            type: string
            description: 借款人姓名
          JKRZJLX:
            type: string
            description: 借款人证件类型
          JKRZJHM:
            type: string
            description: 借款人证件号码
          FKYHMC:
            type: string
            description: 放款银行名称
          FKZH:
            type: string
            description: 放款账户
          FKZHHM:
            type: string
            description: 放款账户户名
          SKYHMC:
            type: string
            description: 收款银行名称
          SKZH:
            type: string
            description: 收款账号
          SKZHHM:
            type: string
            description: 收款账号户名
          DKFFE:
            type: string
            description: 贷款发放额
          DKFFRQ:
            type: string
            description: 贷款发放日期
          CZY:
            type: string
            description: 操作员
          YWWD:
            type: string
            description: 业务网点
  confirmPaymentResponse:
    properties:
      Res:
        type: object
        properties:
          YWLSH:
            type: string
            description: 业务流水号
          JKHTBH:
            type: string
            description: 借款合同编号
          DKZH:
            type: string
            description: 贷款账号
          JKRXM:
            type: string
            description: 借款人姓名
          JKRZJLX:
            type: string
            description: 借款人证件类型
          JKRZJHM:
            type: string
            description: 借款人证件号码
          FKYHMC:
            type: string
            description: 放款银行名称
          FKZH:
            type: string
            description: 放款账户
          FKZHHM:
            type: string
            description: 放款账户户名
          SKYHMC:
            type: string
            description: 收款银行名称
          SKZH:
            type: string
            description: 收款账号
          SKZHHM:
            type: string
            description: 收款账号户名
          DKFFE:
            type: string
            description: 贷款发放额
          DKFFRQ:
            type: string
            description: 贷款发放日期
          managerInformation:
            description: 经办人信息
            type: array
            items:
              type: object
              properties:
                CZY:
                  type: string
                  description: 操作员
                WD_id:
                  type: string
                  description: 网点id
                YWWD:
                  type: string
                  description: 业务网点
...
  CommonResponses:
    properties:
      Res:
        type: object
        properties:
          status:
            description: 操作状态
            type: string
          YWLSH:
            description: 业务流水号(如：若是没有流水号id,那它就是各模块的模块id)
            type: string
  Error:
    properties:
      Res:
        type: object
        properties:
          code:
            type: integer
            description: 操作码
          msg:
            type: string
            description: 操作信息